Add is.vivaldi() check
Added a method to check if the browser is Vivaldi. The method takes one parameter which is optional.
The parameter is to check major version of the browser. There's also 'not' interface option too.

You can checkout Vivaldi here: https://vivaldi.com/ and the latest build's user agent string when this commit
was made here: http://user-agents.me/search?q=Vivaldi+1.0.162.4

Since this is an environment related addition, it's not possible to add tests.